requirements.txt
Django==1.6.8
djangorestframework==3.0.0
